Fix: Separate test code from code under test.
Put DATA_DIR in conftest (test code) and in data_dir (code under test).
