+++ tests/functional/data_dir.py
@@ -15,11 +15,11 @@
 import sys
 import os.path
 
-# Global variables
-# ----------------
-_FILE_DIR = os.path.dirname(os.path.abspath(__file__))
-# Return the subdirectory data/ of the current file's directory.
-DATA_DIR = os.path.join(_FILE_DIR, 'data')
+# Globals
+# -------
+# Directory storing test-specific data.
+DATA_DIR = os.path.join(os.path.dirname(os.path.abspath(__file__)), 'data')
+
 
 # Functions
 # ---------
